Back/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic==2.9.2
werkzeug==3.0.4
